+++ src/main/java/contracts/repository/ContractRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import contracts.domain.Contract;
 
 @Repository
-public interface ContractRepository extends JpaRepository<Contract, Long> {
+public interface ContractRepository extends JpaRepository<Contract, Integer> {
 
 	@Query(value = "SELECT * FROM CONTRACT WHERE LOWER(Agreement_Title) LIKE CONCAT(LOWER(:search), '%') OR RequestID LIKE CONCAT(:search, '%')", nativeQuery = true)
     public List<Contract> searchContracts(@Param("search") String search);
