+++ src/main/java/contracts/service/ContractService.java
@@ -73,7 +73,7 @@ public class ContractService implements IContractService{
 	}
 	
 	@Override
-	public void updateDetails(Long requestid, User user, List<Status> statusList, String agreement_title, 
+	public void updateDetails(Integer requestid, User user, List<Status> statusList, String agreement_title, 
 			String agreement_type, String description, String agreement_location, String language, 
 			String region, String related_agreements) {
